hook up progress to the dbus

+++ Docky.DBus/Docky.DBus/DockManagerDBusItem.cs
@@ -222,7 +222,7 @@ namespace Docky.DBus
 				} else if (key == "badge") {
 					owner.SetRemoteBadgeText ((string) dict [key]);
 				} else if (key == "progress") {
-					owner.Progress = (int) dict [key];
+					owner.Progress = (double) dict [key];
 				} else if (key == "icon-file") {
 					if (owner is IconDockItem)
 						(owner as IconDockItem).SetRemoteIcon ((string) dict [key]);
